Refactor the apply command into a separate file.
