+++ docs/continuous_delivery.md
@@ -4,6 +4,8 @@
 
 How to use hydros to continuously deliver a set of resources to a kubernetes cluster.
 
-
-
 The repo configuration is used to continuously apply a set of resources in a github repository.
+
+< TODO(jeremy): This document is incomplete>
+
+
